Add transaction to downvotes

+++ proto/database/posts.php
@@ -46,37 +46,33 @@ function upvotePost($postId, $userId)
 	}
 
 	$conn->commit();
-    return $ret; //TODO nao sei para que isto serve :)
+    return $ret;
 }
 
 function downvotePost($postId, $userId)
 {
-    //WIP
     global $conn;
+	$ret = false;
+
+	$conn->beginTransaction();
+
+	$stmt = $conn->prepare("SET TRANSACTION ISOLATION LEVEL SERIALIZABLE READ WRITE");
+	$stmt->execute();
+
+	$stmt = $conn->prepare("DELETE FROM activity WHERE post_id = ? AND user_id = ? AND action = ?");
+	$stmt->execute(array($postId, $userId, 'Upvote'));
+
+	$stmt = $conn->prepare("SELECT * FROM activity WHERE post_id = ? AND user_id = ? AND action = ?");
+	$stmt->execute(array($postId, $userId, 'Downvote'));
+	$results = $stmt->fetchAll();
+
+	if (count($results) == 0) {
+		$stmt = $conn->prepare("INSERT INTO activity (post_id, user_id, action) VALUES (?, ?, ?)");
+		$stmt->execute(array($postId, $userId, 'Downvote'));
+		$ret = true;
+	}
         
-	$stmt = $conn->prepare("
-	BEGIN;
-	SET TRANSACTION 
-	ISOLATION LEVEL SERIALIZABLE
-	READ WRITE;
-	
-	DELETE FROM activity
-	WHERE post_id = ?
-	AND user_id = ?
-	AND action = ?;
-	
-	INSERT INTO activity (post_id, user_id, action) 
-	SELECT ?, ?, ?
-	WHERE NOT EXISTS(
-		SELECT post_id, user_id, action
-		FROM Activity
-		WHERE post_id = ?
-		AND user_id = ?
-		AND action = ?
-	);
-	COMMIT;
-	");
-	$stmt->execute(array($postId, $userId, 'Upvote', $postId, $userId, 'Downvote', $postId, $userId, 'Downvote'));
-    return true; //TODO nao sei para que isto serve :)
+	$conn->commit();
+	return $ret;
 }
 ?>
